Begin deprecation of dynamodb2 to replace dynamodb
