+++ README.rst
@@ -254,47 +254,6 @@ Duplicating a single instance sample would look like this
 
     print(template.to_json())
 
-New DynamoDB module
-===================
-
-The original DynamoDB module did not use a consistent interface with the rest
-of troposphere, and so is being phased out. In order to prepare for this,
-you'll need to make a few changes to DynamoDB resources. The biggest change
-is that many of the DynamoDB properties are now treated like regular properties
-in troposphere rather than as helper functions. For a full example of the
-changes, you can check out this diff between the DynamoDB_Table example using
-the old and the new modules::
-
-  # diff examples/DynamoDB_Table.py examples/DynamoDB2_Table.py
-  5,7c5,7
-  < from troposphere.dynamodb import (Key, AttributeDefinition,
-  <                                   ProvisionedThroughput)
-  < from troposphere.dynamodb import Table
-  ---
-  > from troposphere.dynamodb2 import (KeySchema, AttributeDefinition,
-  >                                    ProvisionedThroughput)
-  > from troposphere.dynamodb2 import Table
-  58c58,61
-  <         AttributeDefinition(Ref(hashkeyname), Ref(hashkeytype)),
-  ---
-  >         AttributeDefinition(
-  >             AttributeName=Ref(hashkeyname),
-  >             AttributeType=Ref(hashkeytype)
-  >         ),
-  61c64,67
-  <         Key(Ref(hashkeyname), "HASH")
-  ---
-  >         KeySchema(
-  >             AttributeName=Ref(hashkeyname),
-  >             KeyType="HASH"
-  >         )
-  64,65c70,71
-  <         Ref(readunits),
-  <         Ref(writeunits)
-  ---
-  >         ReadCapacityUnits=Ref(readunits),
-  >         WriteCapacityUnits=Ref(writeunits)
-
 Community
 =========
